requirements.txt
streamlit
pandas
numpy
scikit-learn
xgboost
shap
matplotlib
joblib
google-genai

